backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]